// sis-ng/src/app/model/personalDetails.ts
/**
 * Smart Insurance System
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';

export interface PersonalDetails { 
    id?: number;
    fullName?: string;
    address?: Address;
    dob?: string;
    relationshipStatus?: PersonalDetails.RelationshipStatusEnum;
    homeOwner?: boolean;
    /**
     * Number of dependents under the age of 16
     */
    dependents?: number;
    employmentStatus?: PersonalDetails.EmploymentStatusEnum;
}
export namespace PersonalDetails {
    export type RelationshipStatusEnum = 'Single' | 'Married' | 'null';
    export const RelationshipStatusEnum = {
        Single: 'Single' as RelationshipStatusEnum,
        Married: 'Married' as RelationshipStatusEnum,
        Null: 'null' as RelationshipStatusEnum
    };
    export type EmploymentStatusEnum = 'FullTime' | 'PartTime' | 'Unemployed' | 'Retired' | 'Student' | 'null';
    export const EmploymentStatusEnum = {
        FullTime: 'FullTime' as EmploymentStatusEnum,
        PartTime: 'PartTime' as EmploymentStatusEnum,
        Unemployed: 'Unemployed' as EmploymentStatusEnum,
        Retired: 'Retired' as EmploymentStatusEnum,
        Student: 'Student' as EmploymentStatusEnum,
        Null: 'null' as EmploymentStatusEnum
    };
}